+++ source/encoder_video_h264_nvenc/plugin.py
@@ -55,6 +55,8 @@ class Settings(PluginSettings):
                                  "-aq-strength:v 8\n"
                                  "-a53cc 0\n"
                                  "-b:v:0 4M\n",
+        "keep_container":        True,
+        "dest_container":        "mkv",
     }
 
     def __init__(self):
@@ -73,6 +75,10 @@ class Settings(PluginSettings):
             "main_options":          self.__set_main_options_form_settings(),
             "advanced_options":      self.__set_advanced_options_form_settings(),
             "custom_options":        self.__set_custom_options_form_settings(),
+            "keep_container":        {
+                "label": "Keep the same container",
+            },
+            "dest_container":        self.__set_destination_container(),
         }
 
     def __set_hw_decoding_checkbox_form_settings(self):
@@ -249,6 +255,33 @@ class Settings(PluginSettings):
             values["display"] = 'hidden'
         return values
 
+    def __set_destination_container(self):
+        values = {
+            "label":      "Set the output container",
+            "input_type":     "select",
+            "select_options": [
+                {
+                    'value': "mkv",
+                    'label': ".mkv - Matroska",
+                },
+                {
+                    'value': "avi",
+                    'label': ".avi - AVI (Audio Video Interleaved)",
+                },
+                {
+                    'value': "mov",
+                    'label': ".mov - QuickTime / MOV",
+                },
+                {
+                    'value': "mp4",
+                    'label': ".mp4 - MP4 (MPEG-4 Part 14)",
+                },
+            ],
+        }
+        if self.get_setting('keep_container'):
+            values["display"] = 'hidden'
+        return values
+
 
 class PluginStreamMapper(StreamMapper):
     def __init__(self):
@@ -382,7 +415,7 @@ def on_worker_process(data):
                 data['exec_command'] += ['-threads', '1']
 
         # Add file in
-        data['exec_command'] += ['-i', data.get('file_in')]
+        data['exec_command'] += ['-i', abspath]
 
         if settings.get_setting('advanced'):
             data['exec_command'] += settings.get_setting('advanced_options').split()
@@ -393,10 +426,14 @@ def on_worker_process(data):
         data['exec_command'] += mapper.get_stream_mapping()
         data['exec_command'] += mapper.get_stream_encoding()
 
-        # Do not remux the file. Keep the file out in the same container
-        split_file_in = os.path.splitext(data.get('file_in'))
         split_file_out = os.path.splitext(data.get('file_out'))
-        data['file_out'] = "{}{}".format(split_file_out[0], split_file_in[1])
+        if settings.get_setting('keep_container'):
+            # Do not remux the file. Keep the file out in the same container
+            split_file_in = os.path.splitext(abspath)
+            data['file_out'] = "{}{}".format(split_file_out[0], split_file_in[1])
+        else:
+            # Force the remux to the configured container
+            data['file_out'] = "{}.{}".format(split_file_out[0], settings.get_setting('dest_container'))
 
         data['exec_command'] += ['-y', data.get('file_out')]