+++ source/encoder_video_h264_nvenc/changelog.md
@@ -1,3 +1,6 @@
+**<span style="color:#56adda">0.0.3</span>**
+- Add ability to set the output container
+
 **<span style="color:#56adda">0.0.2</span>**
 - Refactor for Unmanic v1 PluginHandler compatibility